chore: Refactor program to implement a number guessing game
The program has been refactored to implement a number guessing game. Instead of simply printing a greeting message, the program now generates a random number and prompts the user to guess it. The user's guess is compared to the generated number, and appropriate feedback is provided. The game continues until the user guesses the correct number. This change improves the functionality and interactivity of the program.