feat: Add MakeEntry method to Entry class
The MakeEntry method has been added to the Entry class. This method allows for the creation of a journal entry by setting the date, generating a random question, and adding the user's response. This enhances the functionality of the program by providing a streamlined way to make entries in the journal.